requirements.txt
asgiref==3.8.1
Django==5.1.3
pillow==11.0.0
PyJWT==2.10.1
sqlparse==0.5.2
tzdata==2024.2
